Move store to its own file
